# bachelor/third year/Web_and_Software_Architecture/src/doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText API
  description: API for WASAText messaging platform
  version: 1.0.0
...
security: # Apply the security globally to all operations
  - bearerAuth: [] # use the same name as above
paths:
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}/replyMessage:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    post:
      tags:
        - message
      summary: Reply to a message in a conversation.
      description: Create a new message that replies to an existing message in a conversation.
      operationId: replyMessage
      requestBody:
        description: Message details
        content:
          application/json:
            schema:
              description: "Schema for reply message"
              type: object
              properties:
                content:
                  description: Text or photo content.
                  oneOf:
                  - $ref: '#/components/schemas/Text'
                  - $ref: '#/components/schemas/Photo'
        required: true
      responses:
        "201":
          description: Message reply successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          description: Invalid conversation, message id or content message
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Conversation or message not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/conversations/{conversationId}/messages/{messageId}/comments/:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    put:
      tags:
        - message
      summary: Add a comment to a message
      description: Allows the user to add a reaction or comment to a specific message in a conversation.
      operationId: commentMessage
      requestBody:
        description: Comment detail
        content:
          application/json:
            schema:
              description: "Schema for add comment"
              type: object
              properties:
                content:
                  type: string
                  description: The reaction as an emoji
                  example: "👍"
                  minLength: 1
                  maxLength: 2
                  pattern: '^.*?$'
        required: true
      responses:
        '200':
          description: Comment added successufully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid comment format
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Conversation or message not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{userId}/groups/{groupId}/name:
    parameters:
    - $ref: "#/components/parameters/userId"
    - $ref: "#/components/parameters/groupId"
    put:
      tags:
        - group
      summary: Set the name of a group
      description: Allows a user to change the name of an existing group.
      operationId: setGroupName
      requestBody:
        description: New group name
        content:
          application/json:
            schema:
              description: "Schema for set group name"
              type: object
              properties:
                name:
                  description: Group name
                  type: string
                  example: "group123"
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema:
                description: "Schema for set group name response"
                type: object
                properties:
                  groupId:
                    $ref: "#/components/schemas/GroupId"
                  newName:
                    description: New name of the group.
                    type: string
                    example: "Group name"
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
        '400':
          description: Invalid group name
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/photo:
    parameters:
      - $ref: "#/components/parameters/userId"
    put:
      tags:
        - user
      summary: Set the profile photo
      description: Allows the user to update their profile photo.
      operationId: setMyPhoto
      requestBody:
        description: The photo file to upload
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Photo'
        required: true
      responses:
        '200':
          description: Profile photo updated successfully
          content:
            application/json:
              schema:
                description: "Schema for set my photo response"
                type: object
                properties:
                  userId:
                    $ref: '#/components/schemas/UserId'
                  photo:
                    $ref: '#/components/schemas/Photo'
        '400':
          description: Invalid photo
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{userId}/groups/{groupId}/photo:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/groupId"
    put:
      tags:
        - group
      summary: Set the group photo
      description: Allows the user to change the photo of an existing group.
      operationId: setGroupPhoto
      requestBody:
        description: The photo file to upload
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Photo'
        required: true
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                description: "Schema for set group photo response"
                type: object
                properties:
                  groupId:
                    description: Unique group identifer.
                    type: integer
                    example: 1
                  photoUrl:
                    $ref: '#/components/schemas/Photo'
        '400':
          description: Invalid photo
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group not found
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes: # Define the security scheme type (HTTP bearer)
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
  schemas:
    Message:
      title: Message
      description: "Represent a single Message."
      type: object
      properties:
        id:
          description: Unique message identifier.
          type: integer
          example: 1
          readOnly: true
        timestamp: 
          type: string
          format: date-time
          description: Timestamp when the message was sent.
          example: "2023-11-05T14:30:00Z"
        content:
          description: Text or photo content.
          oneOf:
            - $ref: '#/components/schemas/Text'
            - $ref: '#/components/schemas/Photo'
        sender:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: ["received", "read"]
          description: Status of the sent message indicated by checkmarks
          example: "read"
        commments:
          type: array
          minItems: 0
          maxItems: 500
          description: List of comment on the message
          items:
            $ref: '#/components/schemas/Comment'
        type:
          description: Type of the message
          type: string
          enum: ["standard", "forward", "reply"]
          example: "standard"
        replyToMessageId:
          type: integer
          description: The ID of the message this message is replying to.
          example: 12
    Comment:
      title: Comment
      description: "Represet a single Comment."
      type: object
      properties:
        commentId:
          description: Unique comment identifier.
          type: integer
          example: 1
          readOnly: true
        sender:
          $ref: '#/components/schemas/User'
        content:
          type: string
          description: The reaction as an emoji
          example: "👍"
          pattern: '^.*?$'
          minLength: 1
          maxLength: 2
...
    User:
      title: User
      description: "Represents a single User."
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        username:
          $ref: '#/components/schemas/Username'
    Photo:
      description: Base64 encoded image data
      type: string
      format: byte
      pattern: "^[01]+$"
      minLength: 0
      maxLength: 9999999999999999999
    Username:
      description: Username of the user.
      type: string
      example: "user123"
      minLength: 3
      maxLength: 16
      pattern: "^.*?$"
    UserId:
      description: Unique user identifier
      type: integer
      example: 1
      readOnly: true
    GroupId:
      description: Unique group identifier
      type: integer
      example: 1
      readOnly: true
    Text:
      type: string
      description: Text content of the message
      example: "Hello, how are you?"
      minLength: 1
      maxLength: 500
      pattern: '^.*?$'
  parameters:
    userId:
      schema:
        $ref: '#/components/schemas/UserId'
      name: userId
      in: path
      required: true
      description: User id
    conversationId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier of the conversation."
      name: conversationId
      in: path
      required: true
      description: Conversation id
    messageId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier for the message."
      name: messageId
      in: path
      required: true
      description: Message id
...
    groupId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier for the group."
      name: groupId
      in: path
      required: true
      description: Group id
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    InternalServerError:
      description: Internal server error